refactor(moderation): give BanDetails its own file
